mempool: change fee histogram compaction logic
try to put a single item directly left of large lumps,
instead of many items to the right of them

Anecdote: right now, on mainnet, there are lots of ~1 sat/byte txs in the mempool,
but otherwise the mempool is practically empty, so e.g. 1.1 sat/byte tx could get mined fast.
Server (without this diff) is sending a compact histogram of [[89, 127578], [10, 82676], [1, 14238757]].
So what should the client suggest? 10 sat/byte? :/