session: rework statushash calculation.
If a client requested the status of a very busy address (with cold cache:
no precalc yet), we might have gotten a timeout before we stored any
intermediate status calculated. Then, if the client reconnected and
requested the same, we would get stuck in this loop and never make progress.

With this change, we try to store intermediate hashes sooner, so that even if
there is a timeout, if the client reconnects there is less work to be done.