fix: handle correctly the session limiting mechanism
This prevents the DoS attack vector caused by hitting the max number of sessions.
Prior to this change if the max number of sessions was reached, the server would stop listening for new connections until ALL existing connections were closed.
Now, the listening port is closed, the existing connections remain live and when the lower threshold is reached, the listening socket is reopened.