add warmup budget to PaddedRSTransport
Adds a 1024 (unpadded) byte sending budget to the PaddedRSTransport below which
messages are instantly flushed down the socket before the transport will
beginn waiting for the buffer to reach MIN_PACKET_SIZE (1024).
This allows to get the first couple of messages sent quickly when
a wallet connects to improve its ux.